requirements.txt
Flask==1.1.2
SQLAlchemy==1.3.19
psycopg2-binary==2.8.5
PyJWT==1.7.1
redis==3.5.3
